devops(organisation): refactor to model lib
