+++ libs/model/src/index.ts
@@ -1,1 +1,2 @@
 export * from './lib/movie';
+export * from './lib/organisation';
