+++ libs/organization/src/lib/organization/+state/index.ts
@@ -1,3 +1,1 @@
-export * from './organization.model';
 export * from './organization.service';
-export * from './organization.firestore';
